Working on rigid registration, but broken
This is a broken commit while I go debug find_P.

+++ include/cpd/find_P.hpp
@@ -11,7 +11,7 @@ double find_P(
         const arma::mat& X,
         const arma::mat& Y,
         double sigma2,
-        float outliers,
+        double outliers,
         arma::vec& P1,
         arma::vec& Pt1,
         arma::mat& PX,
