+++ include/cpd/registration/base.hpp
@@ -17,9 +17,9 @@ class Base
 public:
 
     explicit Base(
-            float tol = 1e-5,
+            double tol = 1e-5,
             int max_it = 150,
-            float outliers = 0.1,
+            double outliers = 0.1,
             bool use_fgt = false
             );
 
@@ -31,24 +31,24 @@ public:
             const arma::mat& X,
             const arma::mat& Y,
             double sigma2,
-            float outliers,
+            double outliers,
             arma::vec& P1,
             arma::vec& Pt1,
             arma::mat& PX
             ) const;
 
-    inline float get_tol() const { return m_tol; }
+    inline double get_tol() const { return m_tol; }
     inline int get_max_it() const { return m_max_it; }
-    inline float get_outliers() const { return m_outliers; }
+    inline double get_outliers() const { return m_outliers; }
     inline bool use_fgt() const { return m_use_fgt; }
 
 private:
 
     virtual SpResult execute(const arma::mat& X, const arma::mat& Y) const = 0;
 
-    float m_tol;
+    double m_tol;
     int m_max_it;
-    float m_outliers;
+    double m_outliers;
     bool m_use_fgt;
 
 };
